Moved around munge and saving pipeline metadata to the header. They are both no longer stages.
